backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
channels>=4.0.0
daphne>=4.0.0
requests>=2.28.1
aiohttp>=3.8.3
python-dotenv>=1.0.0